+++ src/lib/components/index.ts
@@ -78,5 +78,5 @@ export { default as ImagePreview } from './imagePreview.svelte';
 export { default as MfaChallengeFormList } from './mfaChallengeFormList.svelte';
 export { default as BottomModalAlert } from './bottomModalAlert.svelte';
 export { default as Navbar } from './navbar.svelte';
-export { default as BreadcrumbsConsole } from './breadcrumbsConsole.svelte';
+export { default as Breadcrumbs } from './breadcrumbs.svelte';
 export { default as Sidebar } from './sidebar.svelte';
